New hashing, rearaange search etc.

+++ ChessEngine/ChessEngine_lib/move_search/search_params.h
@@ -5,11 +5,11 @@
 const int POS_INF_CHESS = 1000000;
 const int NEG_INF_CHESS = -POS_INF_CHESS;
 const int MATE_SCORE = POS_INF_CHESS / 10;
-const int MAX_DEPTH = 100;
+const short MAX_DEPTH = 100;
 const int MAX_MATE_DEPTH = 10 * MAX_DEPTH;
 
 struct BestMoveSearchParameters {
-	int depth;
+	short depth;
 	int time_limit;
 };
