+++ ChessEngine/ChessEngine_lib/move_search/transposition_table.h
@@ -14,7 +14,7 @@ enum TranspositionTableEntrySearchType {
 
 struct TranspositionTableEntry {
 	zobrist_hash zobrist_hash;
-	int depth;
+	short depth;
 	int value;
 	TranspositionTableEntryNodeType node_type;
 	Move best_move;
@@ -33,7 +33,7 @@ public:
 	int correct_mate_for_retrieval(int score, int ply);
 	int count_entries();
 	bool key_in_table(zobrist_hash hash);
-	void put(zobrist_hash hash, int depth, int score, int ply, Move best_move, TranspositionTableEntryNodeType node_type);
+	void put(zobrist_hash hash, short depth, int score, int ply, Move best_move, TranspositionTableEntryNodeType node_type);
 	TranspositionTableSearchResults probe_for_move_ordering(zobrist_hash hash);
 	TranspositionTableSearchResults probe_for_search(zobrist_hash hash, int depth, int ply);
 	explicit TranspositionTable(uint64_t size = 3'000'000);
